# Conflicts:
#	nekoyume/Assets/AddressableAssets/UI/Module/CellView/GuidedQuestCell.prefab
#	nekoyume/Assets/AddressableAssets/UI/Module/GuidedQuest.prefab
#	nekoyume/Assets/Resources/UI/Prefabs/UI_Menu.prefab
#	nekoyume/Assets/_Scripts/Game/Stage.cs
#	nekoyume/Assets/_Scripts/UI/Scroller/GuidedQuestCell.cs

+++ nekoyume/Assets/_Scripts/UI/Model/ItemCountAndPricePopup.cs
@@ -1,17 +1,18 @@
+using System.Numerics;
 using UniRx;
 
 namespace Nekoyume.UI.Model
 {
     public class ItemCountAndPricePopup : ItemCountPopup<ItemCountAndPricePopup>
     {
-        public readonly ReactiveProperty<decimal> Price = new ReactiveProperty<decimal>(10);
+        public readonly ReactiveProperty<BigInteger> Price = new ReactiveProperty<BigInteger>(10);
         public readonly ReactiveProperty<bool> PriceInteractable = new ReactiveProperty<bool>(true);
-        
+
         public override void Dispose()
         {
             Price.Dispose();
             PriceInteractable.Dispose();
-            
+
             base.Dispose();
         }
     }
